insertion sort logic
